Moved GeometryUDT to provide substitute for the new geomesa version. Added packaging of same to build.
