+++ pyrasterframes/python/pyrasterframes/__init__.py
@@ -45,16 +45,6 @@ def _convertDF(df, sp_key = None, metadata = None):
             df._jdf, _to_java_column(sp_key), json.dumps(metadata)), ctx._spark_session)
 
 
-_prevFJ = UserDefinedType.fromJson
-# this classmethod annotation is needed for 2.7
-@classmethod
-def _fromJson(cls, json_val):
-    if str(json_val['class']).startswith('org.apache.spark.sql.jts'):
-        json_val['pyClass'] = 'pyrasterframes.GeometryUDT'
-
-    return _prevFJ(json_val)
-
-
 # Patch new method on SparkSession to mirror Scala approach
 SparkSession.withRasterFrames = _rf_init
 
@@ -65,9 +55,6 @@ DataFrame.asRF = _convertDF
 # TODO: make sure this supports **options
 DataFrameReader.geotiff = lambda df_reader, path: _reader(df_reader, "geotiff", path)
 DataFrameReader.geotrellis = lambda df_reader, path: _reader(df_reader, "geotrellis", path)
-
-# If you don't have Python support, you will get it anyway
-UserDefinedType.fromJson = _fromJson
