+++ pyrasterframes/python/pyrasterframes/types.py
@@ -13,7 +13,7 @@ from pyspark.ml.wrapper import JavaTransformer
 from pyspark.ml.util import JavaMLReadable, JavaMLWritable
 from pyrasterframes.context import _checked_context
 
-__all__ = ['RFContext', 'RasterFrame', 'TileUDT', 'GeometryUDT', 'TileExploder']
+__all__ = ['RFContext', 'RasterFrame', 'TileUDT', 'TileExploder']
 
 class RFContext(object):
     """
@@ -106,33 +106,6 @@ class TileUDT(UserDefinedType):
 
     def deserialize(self, datum):
         return _checked_context().generateTile(datum[0], datum[1], datum[2], datum[3])
-
-
-
-class GeometryUDT(UserDefinedType):
-    """User-defined type (UDT).
-
-    .. note:: WARN: Internal use only.
-    """
-
-    @classmethod
-    def sqlType(self):
-        return StructField("wkb", BinaryType(), False)
-
-    @classmethod
-    def module(cls):
-        return 'pyrasterframes'
-
-    @classmethod
-    def scalaUDT(cls):
-        return 'org.apache.spark.sql.jts.GeometryUDT'
-
-    def serialize(self, obj):
-        if (obj is None): return None
-        return Row(obj.toBytes)
-
-    def deserialize(self, datum):
-        return _checked_context().generateGeometry(datum[0])
